Refactor: user message tool type
